LOTS more fixes
Made the lobby player and game play inheirit from the same script. They should be more consistent and there should be footstep + player sounds in the lobby.

Created the TestKit which will use the in game camera, but no sound. Currently, it does not spawn in network objects so networked things like the jump pad, weapons, and sugar collected won't work (yet). Put the prefab in the level and a player will spawn in.